+++ packages/curvenote-cli/src/sync/init/types.ts
@@ -1,0 +1,9 @@
+export type Options = {
+  branch?: string;
+  force?: boolean;
+  yes?: boolean;
+  domain?: string;
+  writeTOC?: boolean;
+};
+
+export const CURVENOTE_YML = 'curvenote.yml';
